Refactored. Added init_config
